+++ Sources/KeyboardKit/_Keyboard/View+KeyboardButton.swift
@@ -12,12 +12,15 @@ public extension View {
 
     /**
      Apply keyboard button style and gestures to the view.
+     
+     The optional `edgeInsets` parameter can be used to add 
 
      - Parameters:
        - action: The keyboard action to trigger.
        - style: The keyboard style to apply.
        - actionHandler: The keyboard action handler to use.
        - calloutContext: The callout context to affect, if any.
+       - edgeInsets: The edge insets to apply to the interactable area, if any.
        - isPressed: An optional binding that can be used to observe the button pressed state.
        - isInScrollView: Whether or not the gestures are used in a scroll view, by default `false`.
        - releaseOutsideTolerance: The percentage of the button size that should span outside the button bounds and still count as a release, by default `0.75`.
@@ -27,15 +30,15 @@ public extension View {
         style: KeyboardStyle.Button,
         actionHandler: KeyboardActionHandler,
         calloutContext: CalloutContext?,
+        edgeInsets: EdgeInsets = .init(),
         isPressed: Binding<Bool> = .constant(false),
         isInScrollView: Bool = false,
         releaseOutsideTolerance: Double = 1
     ) -> some View {
         self
-            .keyboardButtonStyle(
-                style,
-                isPressed: isPressed.wrappedValue
-            )
+            .keyboardButtonStyle(style)
+            .padding(edgeInsets)
+            .contentShape(Rectangle())
             .keyboardButtonGestures(
                 for: action,
                 actionHandler: actionHandler,
@@ -55,21 +58,28 @@ struct View_KeyboardButton_Previews: PreviewProvider {
         var isPressed = false
         
         var body: some View {
-            VStack(spacing: 20) {
-                Text("Is pressed: \(isPressed.description)")
-                button(for: Text("a"), style: .preview1)
-                button(for: Text("A"), style: .preview2)
-                button(for: Image.keyboardGlobe, style: .preview1)
+            VStack {
+                VStack(spacing: 20) {
+                    button(for: Text("a"), style: .preview1)
+                    button(
+                        for: Text("A"),
+                        style: .preview2,
+                        insets: .init(top: 5, leading: 10, bottom: 15, trailing: 20)
+                    )
+                    button(for: Image.keyboardGlobe, style: .preview1)
+                }
+                .padding()
+                .background(Color.gray.opacity(0.5))
+                .cornerRadius(20)
+                
+                Text("\(isPressed ? "Pressed": "Not Pressed")")
             }
-            .padding()
-            .background(Color.gray)
-            .cornerRadius(10)
-            .environment(\.sizeCategory, .extraExtraLarge)
         }
         
         func button<Content: View>(
             for content: Content,
-            style: KeyboardStyle.Button
+            style: KeyboardStyle.Button,
+            insets: EdgeInsets = .init()
         ) -> some View {
             content
                 .padding()
@@ -78,6 +88,7 @@ struct View_KeyboardButton_Previews: PreviewProvider {
                     style: style,
                     actionHandler: .preview,
                     calloutContext: .preview,
+                    edgeInsets: insets,
                     isPressed: $isPressed
                 )
         }